+++ CHANGELOG.md
@@ -1,3 +1,20 @@
+# 0.6.7
+## Additions
+- Chamfer distance introduced in https://arxiv.org/pdf/1612.00603 for dense batches
+
+# 0.6.6
+## Additions
+- Windows support
+
+
+## Change
+- Develop with python 3.7
+
+## Bug fix
+- Fixed bug in region growing related to batching
+- Ball query for partial dense data on GPU was returning only the first point. Fixed now
+
+
 # 0.6.5
 
 ## Additions